requirements.txt
cog==0.9.12
decord==0.6.0
einops==0.3.0
fairscale==0.4.13
fsspec==2023.9.0
ftfy==6.2.0
horovod==0.28.1
huggingface_hub==0.23.0
matplotlib==3.7.2
mmcv==2.0.1
mmengine==0.8.4
mmsegmentation==1.1.1
nltk==3.8.1
numpy==2.0.0
opencv_python==4.6.0.66
opencv_python_headless==4.8.0.76
openpyxl==3.1.2
Pillow==10.4.0
pycocoevalcap==1.2
pycocotools==2.0.7
regex==2023.8.8
ruamel_yaml==0.18.6
safetensors==0.4.3
timm==0.4.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.2
transformers==4.37.2
